+++ docs/release_notes.rst
@@ -14,6 +14,7 @@ Release notes
 * New suppression step added to accept at most three pixels in a cluster, with the condition
   that all the pixels are neighbors.
 * Pull requests merged and issues closed:
+
   - https://github.com/lucabaldini/hexsample/pull/9
   - https://github.com/lucabaldini/hexsample/pull/94
   - https://github.com/lucabaldini/hexsample/pull/92